refactor: update MessageChannel file name casing
